completed 1.0.0 version

+++ app/src/main/java/com/pkm/pinme/data/remote/response/FilterResponse.kt
@@ -30,7 +30,7 @@ data class ArItem(
 	val id: String? = null,
 
 	@field:SerializedName("positionX")
-	val positionX: Int? = null,
+	val positionX: Float? = null,
 
 ) : Parcelable
